+++ config/log_config.json
@@ -1,7 +1,7 @@
 {
     "appenders": {
-      "log": { "type": "dateFile", "filename": "./logs/log.log" },
-      "error": { "type": "dateFile", "filename": "./logs/error.log" },
+      "log": { "type": "dateFile", "filename": "../logs/log.log" },
+      "error": { "type": "dateFile", "filename": "../logs/error.log" },
       "error-filtered": { "type": "logLevelFilter", "appender": "error", "level": "error" },
       "console": { "type": "console" }
     },
