+++ seed/seed_db.js
@@ -4,7 +4,7 @@ const mkdirp = require("mkdirp");
 const { execSync } = require("child_process");
 const unzipper = require("unzipper")
 const mysql = require("promise-mysql");
-const config = require("../config.json");
+const config = require("../app_config.json");
 const rmfr = require('rmfr');
 const fileUrl = "http://kpop.aoimirai.net/download.php";
 const logger = require("../logger")("seed_db");
