+++ src/commands/end.js
