+++ src/commands/play.js
