+++ src/commands/ransom.js
