+++ src/commands/stop.js
