+++ src/logger.js
@@ -1,5 +1,5 @@
 const log4js = require('log4js');
 module.exports = (name) => {
-    log4js.configure('./log4js.json');
+    log4js.configure('../config/log_config.json');
     return log4js.getLogger(name);
 }
