+++ src/models/game_session.js
