+++ src/models/scoreboard.js
